+++ src/cppsim/observable.cpp
@@ -3,15 +3,10 @@
 #include <stdarg.h>
 #include <stdio.h>
 #include "observable.hpp"
-#include "pauli_operator.hpp"
-#include "type.hpp"
 #include "utility.hpp"
 #include "state.hpp"
-#include <vector>
-#include <string>
 #include <fstream>
 #include <iostream>
-#include <utility>
 
 void HermitianQuantumOperator::add_operator(const PauliOperator* mpt){
     if (std::abs(mpt->get_coef().imag()) > 0){
@@ -36,53 +31,30 @@ CPPCTYPE HermitianQuantumOperator::get_expectation_value(const QuantumStateBase*
 namespace observable{
     HermitianQuantumOperator* create_observable_from_openfermion_file(std::string file_path){
         UINT qubit_count = 0;
+        std::vector<CPPCTYPE> coefs;
+        std::vector<std::string> ops;
 
         std::ifstream ifs;
         ifs.open(file_path);
-        double coef_real, coef_imag;
+
         if (!ifs){
             std::cerr << "ERROR: Cannot open file" << std::endl;
 			return NULL;
 		}
 
+        // loading lines and check qubit_count
+        double coef_real, coef_imag;
+        std::string str_buf;
+        std::vector<std::string> index_list;
+
         std::string line;
-        std::vector<CPPCTYPE> coefs;
-        std::vector<std::string> ops;
-        char buf[256];
-        char symbol_j[1];
-        UINT matches;
         while (getline(ifs, line)) {
-            std::vector<std::string> index_list;
 
-            if(line[0]=='('){
-                matches = std::sscanf(line.c_str(), "(%lf+%lfj) [%[^]]]", &coef_real, &coef_imag, buf);
-                if (matches < 2){
-                    matches = std::sscanf(line.c_str(), "(%lf-%lfj) [%[^]]]", &coef_real, &coef_imag, buf);
-                    coef_imag = -coef_imag;
-                }
-                if (matches < 3){
-                    std::strcpy(buf, "I0");
-                }
-            }else{
-                matches = std::sscanf(line.c_str(), "%lf%[j] [%[^]]]", &coef_imag, symbol_j, buf);
-                coef_real = 0.;
-                if (matches < 3){
-                    std::strcpy(buf, "I0");
-                }
-                if (symbol_j[0] != 'j'){
-                    matches = std::sscanf(line.c_str(), "%lf [%[^]]]", &coef_real, buf);
-                    coef_imag = 0.;
-                    if (matches < 2){
-                        std::strcpy(buf, "I0");
-                    }
-                }
-                if (matches == 0){
-                    continue;
-                }
-            }
+            std::tuple<double, double, std::string> parsed_items = parse_openfermion_line(line);
+            coef_real = std::get<0>(parsed_items);
+            coef_imag = std::get<1>(parsed_items);
+            str_buf = std::get<2>(parsed_items);
 
-            std::string str_buf(buf, std::strlen(buf));
-            chfmt(str_buf);
             CPPCTYPE coef(coef_real, coef_imag);
             coefs.push_back(coef);
             ops.push_back(str_buf);
@@ -111,48 +83,22 @@ namespace observable{
 
     HermitianQuantumOperator* create_observable_from_openfermion_text(std::string text){
         UINT qubit_count = 0;
-
-        std::vector<std::string> lines;
         std::vector<CPPCTYPE> coefs;
         std::vector<std::string> ops;
+
+        std::vector<std::string> lines;
         double coef_real, coef_imag;
+        std::string str_buf;
+        std::vector<std::string> index_list;
 
         lines = split(text, "\n");
-        char buf[256];
-        char symbol_j[1];
-        UINT matches;
         for (std::string line: lines){
-            std::vector<std::string> index_list;
 
-            if(line[0]=='('){
-                matches = std::sscanf(line.c_str(), "(%lf+%lfj) [%[^]]]", &coef_real, &coef_imag, buf);
-                if (matches < 2){
-                    matches = std::sscanf(line.c_str(), "(%lf-%lfj) [%[^]]]", &coef_real, &coef_imag, buf);
-                    coef_imag = -coef_imag;
-                }
-                if (matches < 3){
-                    std::strcpy(buf, "I0");
-                }
-            }else{
-                matches = std::sscanf(line.c_str(), "%lf%[j] [%[^]]]", &coef_imag, symbol_j, buf);
-                coef_real = 0.;
-                if (matches < 3){
-                    std::strcpy(buf, "I0");
-                }
-                if (symbol_j[0] != 'j'){
-                    matches = std::sscanf(line.c_str(), "%lf [%[^]]]", &coef_real, buf);
-                    coef_imag = 0.;
-                    if (matches < 2){
-                        std::strcpy(buf, "I0");
-                    }
-                }
-                if (matches == 0){
-                    continue;
-                }
-            }
+            std::tuple<double, double, std::string> parsed_items = parse_openfermion_line(line);
+            coef_real = std::get<0>(parsed_items);
+            coef_imag = std::get<1>(parsed_items);
+            str_buf = std::get<2>(parsed_items);
 
-            std::string str_buf(buf, std::strlen(buf));
-            chfmt(str_buf);
             CPPCTYPE coef(coef_real, coef_imag);
             coefs.push_back(coef);
             ops.push_back(str_buf);
@@ -164,6 +110,7 @@ namespace observable{
                     qubit_count = n;
             }
         }
+
         HermitianQuantumOperator* hermitian_quantum_operator = new HermitianQuantumOperator(qubit_count);
 
         for (UINT i = 0; i < ops.size(); ++i){
@@ -176,7 +123,9 @@ namespace observable{
 
     std::pair<HermitianQuantumOperator*, HermitianQuantumOperator*> create_split_observable(std::string file_path){
         UINT qubit_count = 0;
-        double coef_real, coef_imag;
+        std::vector<CPPCTYPE> coefs;
+        std::vector<std::string> ops;
+
         std::ifstream ifs;
         ifs.open(file_path);
 
@@ -186,44 +135,18 @@ namespace observable{
         }
 
         // loading lines and check qubit_count
+        double coef_real, coef_imag;
+        std::string str_buf;
+        std::vector<std::string> index_list;
+
         std::string line;
-        std::vector<CPPCTYPE> coefs;
-        std::vector<std::string> ops;
-        char buf[256];
-        char symbol_j[1];
-        UINT matches;
         while (getline(ifs, line)) {
-            std::vector<std::string> index_list;
 
-            if(line[0]=='('){
-                matches = std::sscanf(line.c_str(), "(%lf+%lfj) [%[^]]]", &coef_real, &coef_imag, buf);
-                if (matches < 2){
-                    matches = std::sscanf(line.c_str(), "(%lf-%lfj) [%[^]]]", &coef_real, &coef_imag, buf);
-                    coef_imag = -coef_imag;
-                }
-                if (matches < 3){
-                    std::strcpy(buf, "I0");
-                }
-            }else{
-                matches = std::sscanf(line.c_str(), "%lf%[j] [%[^]]]", &coef_imag, symbol_j, buf);
-                coef_real = 0.;
-                if (matches < 3){
-                    std::strcpy(buf, "I0");
-                }
-                if (symbol_j[0] != 'j'){
-                    matches = std::sscanf(line.c_str(), "%lf [%[^]]]", &coef_real, buf);
-                    coef_imag = 0.;
-                    if (matches < 2){
-                        std::strcpy(buf, "I0");
-                    }
-                }
-                if (matches == 0){
-                    continue;
-                }
-            }
+            std::tuple<double, double, std::string> parsed_items = parse_openfermion_line(line);
+            coef_real = std::get<0>(parsed_items);
+            coef_imag = std::get<1>(parsed_items);
+            str_buf = std::get<2>(parsed_items);
 
-            std::string str_buf(buf, std::strlen(buf));
-            chfmt(str_buf);
             CPPCTYPE coef(coef_real, coef_imag);
             coefs.push_back(coef);
             ops.push_back(str_buf);
