+++ src/cppsim/pauli_operator.cpp
@@ -117,296 +117,3 @@ PauliOperator* PauliOperator::copy() const {
 }
 
 
-bool check_Pauli_operator(const GeneralQuantumOperator* observable, const PauliOperator* pauli_operator) {
-	auto vec = pauli_operator->get_index_list();
-	UINT val = 0;
-	if (vec.size() > 0) {
-		val = std::max(val, *std::max_element(vec.begin(), vec.end()));
-	}
-	return val < (observable->get_qubit_count());
-}
-
-GeneralQuantumOperator::GeneralQuantumOperator(UINT qubit_count){
-    _qubit_count = qubit_count;
-    _is_hermitian = true;
-}
-
-GeneralQuantumOperator::~GeneralQuantumOperator(){
-    for(auto& term : this->_operator_list){
-        delete term;
-    }
-}
-
-void GeneralQuantumOperator::add_operator(const PauliOperator* mpt){
-    PauliOperator* _mpt = mpt->copy();
-	if (!check_Pauli_operator(this, _mpt)) {
-		std::cerr << "Error: GeneralQuantumOperator::add_operator(const PauliOperator*): pauli_operator applies target qubit of which the index is larger than qubit_count" << std::endl;
-		return;
-	}
-    if (this->_is_hermitian && std::abs(_mpt->get_coef().imag()) > 0){
-        this->_is_hermitian = false;
-    }
-    this->_operator_list.push_back(_mpt);
-}
-
-void GeneralQuantumOperator::add_operator(CPPCTYPE coef, std::string pauli_string) {
-	PauliOperator* _mpt = new PauliOperator(pauli_string, coef);
-	if (!check_Pauli_operator(this, _mpt)) {
-		std::cerr << "Error: GeneralQuantumOperator::add_operator(double,std::string): pauli_operator applies target qubit of which the index is larger than qubit_count" << std::endl;
-		return;
-	}
-    if (this->_is_hermitian && std::abs(coef.imag()) > 0){
-        this->_is_hermitian = false;
-    }
-	this->add_operator(_mpt);
-}
-
-CPPCTYPE GeneralQuantumOperator::get_expectation_value(const QuantumStateBase* state) const {
-	if (this->_qubit_count != state->qubit_count) {
-		std::cerr << "Error: GeneralQuantumOperator::get_expectation_value(const QuantumStateBase*): invalid qubit count" << std::endl;
-		return 0.;
-	}
-    CPPCTYPE sum = 0;
-    for (auto pauli : this->_operator_list) {
-        sum += pauli->get_expectation_value(state);
-    }
-    return sum;
-}
-
-CPPCTYPE GeneralQuantumOperator::get_transition_amplitude(const QuantumStateBase* state_bra, const QuantumStateBase* state_ket) const {
-	if (this->_qubit_count != state_bra->qubit_count || this->_qubit_count != state_ket->qubit_count) {
-		std::cerr << "Error: GeneralQuantumOperator::get_transition_amplitude(const QuantumStateBase*, const QuantumStateBase*): invalid qubit count" << std::endl;
-		return 0.;
-	}
-
-	CPPCTYPE sum = 0;
-    for (auto pauli : this->_operator_list) {
-        sum += pauli->get_transition_amplitude(state_bra, state_ket);
-    }
-    return sum;
-}
-
-namespace quantum_operator{
-    GeneralQuantumOperator* create_general_quantum_operator_from_openfermion_file(std::string file_path){
-        UINT qubit_count = 0;
-
-        std::ifstream ifs;
-        ifs.open(file_path);
-        double coef_real, coef_imag;
-        if (!ifs){
-            std::cerr << "ERROR: Cannot open file" << std::endl;
-			return NULL;
-		}
-
-        std::string line;
-        std::vector<CPPCTYPE> coefs;
-        std::vector<std::string> ops;
-        char buf[256];
-        char symbol_j[1];
-        UINT matches;
-        while (getline(ifs, line)) {
-            std::vector<std::string> index_list;
-
-            if(line[0]=='('){
-                matches = std::sscanf(line.c_str(), "(%lf+%lfj) [%[^]]]", &coef_real, &coef_imag, buf);
-                if (matches < 2){
-                    matches = std::sscanf(line.c_str(), "(%lf-%lfj) [%[^]]]", &coef_real, &coef_imag, buf);
-                    coef_imag = -coef_imag;
-                }
-                if (matches < 3){
-                    std::strcpy(buf, "I0");
-                }
-            }else{
-                matches = std::sscanf(line.c_str(), "%lf%[j] [%[^]]]", &coef_imag, symbol_j, buf);
-                coef_real = 0.;
-                if (matches < 3){
-                    std::strcpy(buf, "I0");
-                }
-                if (symbol_j[0] != 'j'){
-                    matches = std::sscanf(line.c_str(), "%lf [%[^]]]", &coef_real, buf);
-                    coef_imag = 0.;
-                    if (matches < 2){
-                        std::strcpy(buf, "I0");
-                    }
-                }
-                if (matches == 0){
-                    continue;
-                }
-            }
-
-            std::string str_buf(buf, std::strlen(buf));
-            chfmt(str_buf);
-            CPPCTYPE coef(coef_real, coef_imag);
-            coefs.push_back(coef);
-            ops.push_back(str_buf);
-            index_list = split(str_buf, "IXYZ ");
-
-            for (UINT i = 0; i < index_list.size(); ++i){
-                UINT n = std::stoi(index_list[i]) + 1;
-                if (qubit_count < n)
-                    qubit_count = n;
-            }
-        }
-        if (!ifs.eof()){
-            std::cerr << "ERROR: Invalid format" << std::endl;
-			return NULL;
-		}
-        ifs.close();
-
-        GeneralQuantumOperator* general_quantum_operator = new GeneralQuantumOperator(qubit_count);
-
-        for (UINT i = 0; i < ops.size(); ++i){
-            general_quantum_operator->add_operator(new PauliOperator(ops[i].c_str(), coefs[i]));
-        }
-
-        return general_quantum_operator;
-    }
-
-    GeneralQuantumOperator* create_general_quantum_operator_from_openfermion_text(std::string text){
-        UINT qubit_count = 0;
-
-        std::vector<std::string> lines;
-        std::vector<CPPCTYPE> coefs;
-        std::vector<std::string> ops;
-        double coef_real, coef_imag;
-
-        lines = split(text, "\n");
-        char buf[256];
-        char symbol_j[1];
-        UINT matches;
-        for (std::string line: lines){
-            std::vector<std::string> index_list;
-
-            if(line[0]=='('){
-                matches = std::sscanf(line.c_str(), "(%lf+%lfj) [%[^]]]", &coef_real, &coef_imag, buf);
-                if (matches < 2){
-                    matches = std::sscanf(line.c_str(), "(%lf-%lfj) [%[^]]]", &coef_real, &coef_imag, buf);
-                    coef_imag = -coef_imag;
-                }
-                if (matches < 3){
-                    std::strcpy(buf, "I0");
-                }
-            }else{
-                matches = std::sscanf(line.c_str(), "%lf%[j] [%[^]]]", &coef_imag, symbol_j, buf);
-                coef_real = 0.;
-                if (matches < 3){
-                    std::strcpy(buf, "I0");
-                }
-                if (symbol_j[0] != 'j'){
-                    matches = std::sscanf(line.c_str(), "%lf [%[^]]]", &coef_real, buf);
-                    coef_imag = 0.;
-                    if (matches < 2){
-                        std::strcpy(buf, "I0");
-                    }
-                }
-                if (matches == 0){
-                    continue;
-                }
-            }
-
-            std::string str_buf(buf, std::strlen(buf));
-            chfmt(str_buf);
-            CPPCTYPE coef(coef_real, coef_imag);
-            coefs.push_back(coef);
-            ops.push_back(str_buf);
-            index_list = split(str_buf, "IXYZ ");
-
-            for (UINT i = 0; i < index_list.size(); ++i){
-                UINT n = std::stoi(index_list[i]) + 1;
-                if (qubit_count < n)
-                    qubit_count = n;
-            }
-        }
-        GeneralQuantumOperator* general_quantum_operator = new GeneralQuantumOperator(qubit_count);
-
-        for (UINT i = 0; i < ops.size(); ++i){
-            general_quantum_operator->add_operator(new PauliOperator(ops[i].c_str(), coefs[i]));
-        }
-
-        return general_quantum_operator;
-    }
-
-    std::pair<GeneralQuantumOperator*, GeneralQuantumOperator*> create_split_general_quantum_operator(std::string file_path){
-        UINT qubit_count = 0;
-        std::ifstream ifs;
-        double coef_real, coef_imag;
-
-        ifs.open(file_path);
-
-        if (!ifs){
-            std::cerr << "ERROR: Cannot open file" << std::endl;
-			return std::make_pair((GeneralQuantumOperator*)NULL, (GeneralQuantumOperator*)NULL);
-        }
-
-        // loading lines and check qubit_count
-        std::string line;
-        std::vector<CPPCTYPE> coefs;
-        std::vector<std::string> ops;
-        char buf[256];
-        char symbol_j[1];
-        UINT matches;
-        while (getline(ifs, line)) {
-            std::vector<std::string> index_list;
-
-            if(line[0]=='('){
-                matches = std::sscanf(line.c_str(), "(%lf+%lfj) [%[^]]]", &coef_real, &coef_imag, buf);
-                if (matches < 2){
-                    matches = std::sscanf(line.c_str(), "(%lf-%lfj) [%[^]]]", &coef_real, &coef_imag, buf);
-                    coef_imag = -coef_imag;
-                }
-                if (matches < 3){
-                    std::strcpy(buf, "I0");
-                }
-            }else{
-                matches = std::sscanf(line.c_str(), "%lf%[j] [%[^]]]", &coef_imag, symbol_j, buf);
-                coef_real = 0.;
-                if (matches < 3){
-                    std::strcpy(buf, "I0");
-                }
-                if (symbol_j[0] != 'j'){
-                    matches = std::sscanf(line.c_str(), "%lf [%[^]]]", &coef_real, buf);
-                    coef_imag = 0.;
-                    if (matches < 2){
-                        std::strcpy(buf, "I0");
-                    }
-                }
-                if (matches == 0){
-                    continue;
-                }
-            }
-
-            std::string str_buf(buf, std::strlen(buf));
-            chfmt(str_buf);
-            CPPCTYPE coef(coef_real, coef_imag);
-            coefs.push_back(coef);
-            ops.push_back(str_buf);
-            index_list = split(str_buf, "IXYZ ");
-
-            for (UINT i = 0; i < index_list.size(); ++i){
-                UINT n = std::stoi(index_list[i]) + 1;
-                if (qubit_count < n)
-                    qubit_count = n;
-            }
-        }
-        if (!ifs.eof()){
-            std::cerr << "ERROR: Invalid format" << std::endl;
-			return std::make_pair((GeneralQuantumOperator*)NULL, (GeneralQuantumOperator*)NULL);
-		}
-        ifs.close();
-
-        GeneralQuantumOperator* general_quantum_operator_diag =  new GeneralQuantumOperator(qubit_count);
-        GeneralQuantumOperator* general_quantum_operator_non_diag =  new GeneralQuantumOperator(qubit_count);
-
-        for (UINT i = 0; i < ops.size(); ++i){
-            if (ops[i].find("X") != std::string::npos || ops[i].find("Y") != std::string::npos){
-                general_quantum_operator_non_diag->add_operator(new PauliOperator(ops[i].c_str(), coefs[i]));
-            }else{
-                general_quantum_operator_diag->add_operator(new PauliOperator(ops[i].c_str(), coefs[i]));
-            }
-        }
-
-        return std::make_pair(general_quantum_operator_diag, general_quantum_operator_non_diag);
-    }
-}
-
-
